requirements.txt
flask
opencv-python
tensorflow
keras
pillow
numpy